backend/requirements.txt
setuptools>=69.0.0
# Core backend framework
flask==2.3.2
flask-sqlalchemy==3.0.3
flask-cors==4.0.0
flask-migrate==4.0.4

# Image processing and AI
face_recognition==1.3.0
git+https://github.com/ageitgey/face_recognition_models
opencv-python==4.9.0.80
numpy==1.26.2

# Database and security
bcrypt==4.0.1
flask-jwt-extended==4.6.0

# Utilities
python-dotenv==1.0.0
gunicorn==21.2.0 

# Testing
pytest==7.4.2
pytest-flask==1.2.0

# Reporting 
pandas==2.2.0
xlsxwriter==3.1.9
reportlab==4.0.7